requirements.txt
Flask>=3.0
flask-socketio>=5.3
python-socketio>=5.11
python-engineio>=4.9
gevent>=24.2.1
gevent-websocket>=0.10.1
gunicorn>=21.2
psycopg2-binary>=2.9
Flask-Compress>=1.14
